+++ release-notes.md
@@ -23,6 +23,8 @@ Zay-ES Core v1.6.0 (unrelease)
     indexes for types with IndexedField fields.
 * Added basic support for primitive array types in the SQL layer.
 * Fixed a bug in the SQL layer when saving components with no fields.
+* Updated ComponentTable to automatically add new table fields with 'null' values.
+* Updated FieldTypes.PrimitiveField to support null DB values default to 0.
 
 
 Zay-ES Core v1.5.0 (latest)